chore: rename schemas to typesense-schemas
