+++ server/libs/infra/src/typesense-schemas/album.schema.ts
