+++ server/libs/infra/src/typesense-schemas/asset.schema.ts
