+++ server/libs/infra/src/typesense-schemas/index.ts
